+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use localrep data in CPAlgoViewSet (#763)
 - Local representation of performance assets (#778)
 - Use localrep data in tasks views (#760)
+- Use localrep performances in tasks views (#780)
 
 ### Changed
 - When resyncing the local representation, only fetch the latest events (#656)